requirements.txt
asgiref
Django
django-ckeditor
django-js-asset
django-model-utils
django-mptt
Pillow
sqlparse
whitenoise
